Implemented Function against SQL-Injection and added Repo to Em_Contract Entity

+++ src/main/java/de/saschakiebler/businessmanagement/service/BillService.java
@@ -26,25 +26,34 @@ public class BillService {
 
     public Bill getBillByNumber(String billNumber){
 
-        return billRepository.getBillByNumber(billNumber);
+        if(RecognizeIllegalSQLStatements.recognizeAnySQLRelevantWords(billNumber))
+            return billRepository.getBillByNumber(billNumber);
+
+        return null;
     }
 
     public Bill getBillByID(String bill_id){
+        if (RecognizeIllegalSQLStatements.recognizeAnySQLRelevantWords(bill_id))
+            return billRepository.getBillByID(bill_id);
 
-        return billRepository.getBillByID(bill_id);
+        return null;
     }
 
     public List<Bill> getAllBillsFromClient(String client_id){
-        Client client = clientRepository.getClientByID(client_id);
-        return billRepository.getAllBillsFromClient(client);
+        if (RecognizeIllegalSQLStatements.recognizeAnySQLRelevantWords(client_id)) {
+            Client client = clientRepository.getClientByID(client_id);
+            return billRepository.getAllBillsFromClient(client);
+        }
+        return null;
     }
 
     public void updateBill(Bill bill){
-
+        if (RecognizeIllegalSQLStatements.recognizeAnySQLRelevantWords(bill.toString()))
         billRepository.updateBill(bill);
     }
 
     public void deleteBill(String bill_id){
+        if (RecognizeIllegalSQLStatements.recognizeAnySQLRelevantWords(bill_id))
         billRepository.deleteBill(bill_id);
     }
 }
